MAINT: signal: lint

+++ scipy/signal/_bsplines.py
@@ -9,6 +9,7 @@ from ._spline import cspline2d, sepfir2d
 
 from scipy.special import comb
 from scipy._lib._util import float_factorial
+from scipy.interpolate import BSpline
 
 __all__ = ['spline_filter', 'bspline', 'gauss_spline', 'cubic', 'quadratic',
            'cspline1d', 'qspline1d', 'cspline1d_eval', 'qspline1d_eval']
@@ -141,7 +142,7 @@ def _bspline_piecefunctions(order):
     return funclist, condfuncs
 
 
-msg_bspline="""`scipy.signal.bspline` is deprecated in SciPy 1.11 and will be
+msg_bspline = """`scipy.signal.bspline` is deprecated in SciPy 1.11 and will be
 removed in SciPy 1.13.
 
 The exact equivalent (for a float array `x`) is
@@ -151,6 +152,7 @@ The exact equivalent (for a float array `x`) is
 >>> out = BSpline.basis_element(knots)(x)
 >>> out[(x < knots[0]) | (x > knots[-1])] = 0.0
 """
+
 
 @np.deprecate(message=msg_bspline)
 def bspline(x, n):
@@ -256,7 +258,7 @@ def gauss_spline(x, n):
     return 1 / sqrt(2 * pi * signsq) * exp(-x ** 2 / 2 / signsq)
 
 
-msg_cubic="""`scipy.signal.cubic` is deprecated in SciPy 1.11 and will be
+msg_cubic = """`scipy.signal.cubic` is deprecated in SciPy 1.11 and will be
 removed in SciPy 1.13.
 
 The exact equivalent (for a float array `x`) is
@@ -265,6 +267,7 @@ The exact equivalent (for a float array `x`) is
 >>> out = BSpline.basis_element([-2, -1, 0, 1, 2])(x)
 >>> out[(x < -2 | (x > 2)] = 0.0
 """
+
 
 @np.deprecate(message=msg_cubic)
 def cubic(x):
@@ -307,7 +310,7 @@ def cubic(x):
     True
 
     """
-    ax = abs(asarray(x))
+    ax = abs(asarray(x, dtype=float))
     res = zeros_like(ax)
     cond1 = less(ax, 1)
     if cond1.any():
@@ -319,6 +322,7 @@ def cubic(x):
         res[cond2] = 1.0 / 6 * (2 - ax2) ** 3
     return res
 
+
 def _cubic(x):
     x = asarray(x, dtype=float)
     b = BSpline.basis_element([-2, -1, 0, 1, 2], extrapolate=False)
@@ -327,7 +331,7 @@ def _cubic(x):
     return out
 
 
-msg_quadratic="""`scipy.signal.quadratic` is deprecated in SciPy 1.11 and
+msg_quadratic = """`scipy.signal.quadratic` is deprecated in SciPy 1.11 and
 will be removed in SciPy 1.13.
 
 The exact equivalent (for a float array `x`) is
@@ -336,6 +340,7 @@ The exact equivalent (for a float array `x`) is
 >>> out = BSpline.basis_element([-1.5, -0.5, 0.5, 1.5])(x)
 >>> out[(x < -1.5 | (x > 1.5)] = 0.0
 """
+
 
 @np.deprecate(message=msg_quadratic)
 def quadratic(x):
